+++ ADMF/changelog.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 
+## 1.9.58 (2021-03-04)
+
+- New: Component - Exchange System Objects (at the domain level)
+- Upd: DC Resolution - Added new option to select DC by site, rather than just PDC or random.
+- Upd: Added caching to domain controller resolution so that credential providers may use the same resolution results.
+
 ## 1.8.55 (2021-01-15)
 
 - New: Component - Service Accounts